added a new separate new api/query btn component
