+++ app/client/src/pages/Editor/DataSourceEditor/Connected.tsx
@@ -1,25 +1,16 @@
-import React, { useCallback } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import React from "react";
+import { useSelector } from "react-redux";
 import { useParams } from "react-router";
 import { AppState } from "reducers";
 import { isNil } from "lodash";
-import Button from "components/ads/Button";
 import { getDatasource, getPlugin } from "selectors/entitiesSelector";
 import { Colors } from "constants/Colors";
 import { HeaderIcons } from "icons/HeaderIcons";
 import styled from "styled-components";
-import { createActionRequest } from "actions/actionActions";
-import { createNewApiName, createNewQueryName } from "utils/AppsmithUtils";
-import { getCurrentPageId } from "selectors/editorSelectors";
-import { Action, ApiActionConfig, PluginType } from "entities/Action";
 import { renderDatasourceSection } from "./DatasourceSection";
 import { OnboardingStep } from "constants/OnboardingConstants";
-import { inOnboarding } from "sagas/OnboardingSagas";
 import OnboardingIndicator from "components/editorComponents/Onboarding/Indicator";
-import { Toaster } from "components/ads/Toast";
-import { Variant } from "components/ads/common";
-import { ERROR_ADD_API_INVALID_URL } from "constants/messages";
-import { DEFAULT_API_ACTION_CONFIG } from "constants/ApiEditorConstants";
+import NewActionButton from "./NewActionButton";
 
 const ConnectedText = styled.div`
   color: ${Colors.OXFORD_BLUE};
@@ -46,30 +37,12 @@ const Wrapper = styled.div`
   margin-top: 18px;
 `;
 
-const ActionButton = styled(Button)`
-  padding: 10px 20px;
-  &&&& {
-    height: 36px;
-    //max-width: 120px;
-    width: auto;
-  }
-  span > svg > path {
-    stroke: white;
-  }
-`;
-
 function Connected() {
   const params = useParams<{ datasourceId: string; applicationId: string }>();
   const datasource = useSelector((state: AppState) =>
     getDatasource(state, params.datasourceId),
   );
 
-  // Onboarding
-  const isInOnboarding = useSelector(inOnboarding);
-
-  const dispatch = useDispatch();
-  const actions = useSelector((state: AppState) => state.entities.actions);
-  const currentPageId = useSelector(getCurrentPageId);
   const datasourceFormConfigs = useSelector(
     (state: AppState) => state.entities.plugins.formConfigs,
   );
@@ -77,60 +50,6 @@ function Connected() {
   const plugin = useSelector((state: AppState) =>
     getPlugin(state, datasource?.pluginId ?? ""),
   );
-
-  const createQueryAction = useCallback(() => {
-    const pluginType = plugin?.type;
-    const newActionName =
-      pluginType === PluginType.DB
-        ? createNewQueryName(actions, currentPageId || "")
-        : createNewApiName(actions, currentPageId || "");
-    if (
-      pluginType === PluginType.API &&
-      (!datasource ||
-        !datasource.datasourceConfiguration ||
-        !datasource.datasourceConfiguration.url)
-    ) {
-      Toaster.show({
-        text: ERROR_ADD_API_INVALID_URL(),
-        variant: Variant.danger,
-      });
-      return;
-    }
-
-    const headers = datasource?.datasourceConfiguration?.headers ?? [];
-    const defaultApiActionConfig: ApiActionConfig = {
-      ...DEFAULT_API_ACTION_CONFIG,
-      headers: headers.length ? headers : DEFAULT_API_ACTION_CONFIG.headers,
-    };
-    let payload = {
-      name: newActionName,
-      pageId: currentPageId,
-      pluginId: datasource?.pluginId,
-      datasource: {
-        id: datasource?.id,
-      },
-      actionConfiguration:
-        pluginType === PluginType.API ? defaultApiActionConfig : {},
-      eventData: {
-        actionType: pluginType === PluginType.DB ? "Query" : "API",
-        from: "datasource-pane",
-      },
-    } as Partial<Action>;
-
-    if (datasource)
-      if (isInOnboarding) {
-        // If in onboarding and tooltip is being shown
-        payload = Object.assign({}, payload, {
-          name: "fetch_standup_updates",
-          actionConfiguration: {
-            body:
-              "Select avatar, name, notes from standup_updates order by id desc",
-          },
-        });
-      }
-
-    dispatch(createActionRequest(payload));
-  }, [dispatch, actions, currentPageId, datasource, plugin]);
 
   const currentFormConfig: Array<any> =
     datasourceFormConfigs[datasource?.pluginId ?? ""];
@@ -149,11 +68,10 @@ function Connected() {
         </ConnectedText>
 
         <OnboardingIndicator step={OnboardingStep.EXAMPLE_DATABASE} width={120}>
-          <ActionButton
-            className="t--create-query"
-            icon="plus"
-            onClick={createQueryAction}
-            text={plugin?.type === PluginType.DB ? "New Query" : "New API"}
+          <NewActionButton
+            datasource={datasource}
+            eventFrom="datasource-pane"
+            pluginType={plugin?.type}
           />
         </OnboardingIndicator>
       </Header>
