+++ app/client/src/pages/Editor/IntegrationEditor/ActiveDataSources.tsx
@@ -1,20 +1,13 @@
-import React, { useCallback, useMemo } from "react";
+import React, { useMemo } from "react";
 import styled from "styled-components";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { AppState } from "reducers";
-import { createNewApiName, createNewQueryName } from "utils/AppsmithUtils";
 import { Datasource } from "entities/Datasource";
-import { createActionRequest } from "actions/actionActions";
-import { ApiActionConfig, PluginType } from "entities/Action";
 import DatasourceCard from "./DatasourceCard";
 import Text, { TextType } from "components/ads/Text";
 import Button, { Category, Size } from "components/ads/Button";
 import { thinScrollbar } from "constants/DefaultTheme";
-import { Toaster } from "components/ads/Toast";
-import { Variant } from "components/ads/common";
-import { DEFAULT_API_ACTION_CONFIG } from "constants/ApiEditorConstants";
 import { keyBy } from "lodash";
-import { ERROR_ADD_API_INVALID_URL } from "../../../constants/messages";
 
 const QueryHomePage = styled.div`
   ${thinScrollbar};
@@ -51,7 +44,6 @@ type ActiveDataSourcesProps = {
   dataSources: Datasource[];
   applicationId: string;
   pageId: string;
-  isCreating: boolean;
   location: {
     search: string;
   };
@@ -63,61 +55,12 @@ type ActiveDataSourcesProps = {
 };
 
 function ActiveDataSources(props: ActiveDataSourcesProps) {
-  const { dataSources, isCreating, pageId } = props;
-  const dispatch = useDispatch();
+  const { dataSources } = props;
 
-  const actions = useSelector((state: AppState) => state.entities.actions);
   const plugins = useSelector((state: AppState) => {
     return state.entities.plugins.list;
   });
   const pluginGroups = useMemo(() => keyBy(plugins, "id"), [plugins]);
-
-  const handleCreateNewQuery = useCallback(
-    (dataSource: Datasource, pluginType: PluginType) => {
-      if (
-        pluginType === PluginType.API &&
-        (!dataSource ||
-          !dataSource.datasourceConfiguration ||
-          !dataSource.datasourceConfiguration.url)
-      ) {
-        Toaster.show({
-          text: ERROR_ADD_API_INVALID_URL(),
-          variant: Variant.danger,
-        });
-        return;
-      }
-      if (pageId) {
-        const newActionName =
-          pluginType === PluginType.DB
-            ? createNewQueryName(actions, pageId)
-            : createNewApiName(actions, pageId);
-
-        const headers = dataSource?.datasourceConfiguration?.headers ?? [];
-        const defaultApiActionConfig: ApiActionConfig = {
-          ...DEFAULT_API_ACTION_CONFIG,
-          headers: headers.length ? headers : DEFAULT_API_ACTION_CONFIG.headers,
-        };
-        const payload = {
-          name: newActionName,
-          pageId,
-          datasource: {
-            id: dataSource.id,
-          },
-          eventData: {
-            actionType: pluginType === PluginType.DB ? "Query" : "API",
-            from: "datasources",
-            dataSource: dataSource.name,
-          },
-          pluginId: dataSource.pluginId,
-          actionConfiguration:
-            pluginType === PluginType.API ? defaultApiActionConfig : {},
-        };
-
-        dispatch(createActionRequest(payload));
-      }
-    },
-    [dispatch, actions, pageId],
-  );
 
   if (dataSources.length === 0) {
     return (
@@ -142,9 +85,7 @@ function ActiveDataSources(props: ActiveDataSourcesProps) {
         return (
           <DatasourceCard
             datasource={datasource}
-            isCreating={isCreating}
             key={`${datasource.id}_${idx}`}
-            onCreateQuery={handleCreateNewQuery}
             plugin={pluginGroups[datasource.pluginId]}
           />
         );
